moving subscription management into the API
